Refactor test code to give methods with meaningful names
Given, When, Then is the most popular naming convention for test case.
In this case, since we have only 1 production method so the Given part
does not seem necessary. So we just put the When part (when_*) and Then
part (get_*) in the method's names.

Be aware that ABAP method/class name are limited up to 30 characters,
unfortunately.