Keep refactoring our test code by removing duplicates
There is a duplicate of line in both test methods i.e.:
DATA(lo_fb) = NEW lcl_fizzbuzz01( ).

Then we extract it to a new method with meaningful name
for better readability.

As a rule of thumb, you should avoid using setup and teardown
method. Because their names are so not meaningful and it makes
the reader to scroll through the code i.e. bad readability.