Fix our production code to make 6 passed
Use the same technique we did for 3 and 5 by adding a new condition
check at the top of code. Green bar again now.

I know this is stupid as we all know what we should do to solve this
problem but just let's pretend that we don't know it just for practicing
purpose as it may not be obvious for other problems.

Hold on and we will make it in the next step i.e. refactor