Fix production code to make 15 pass
Use the same old solution by adding a new IF at the top of the code.

Why I don't make it like 3 and 5 right away by using modulus logic?
Well, you can do that but taking too big step may result in insufficient
test cases. If you did it now, the next test case of 30 will not be a
failing test so it is possible that you skip it.

You can see so far that each test scenario should have at least 2 test
case to ensure there's no hardcoding. If you plan to add 30 later then
you may use modulus now. But since we're practicing now so let's follow
the step so we create the right habit. The real problems may not be easy
like this.