Make a big change to our production code by concatenate string
You are totally changing the logic of the program in order to
concatenate string correctly. What if you don't have those tests in
place? How do you ensure your changes do not have negative impact. With
those tests, if you see the green bar, you can be confident that
everything is still working as it should be.

Actually, you can keep refactoring the code but I think I'm satisfied
enough with this now. Exercise 1.1 is done!