Generalize our methods into an interface method
Same method that behave differently is the concept of Polymorphism,
isn't it? Now, let's generalize our methods. There are two ways to
generalize: by inheritance and by interface. Which one give you lower
coupling? Interface, isn't it?

So let's create a new interface here with the same signature as
isFizzRull and isBuzzRule.